Added pagination
I updated the list_em_posts page to also display on multiple pages depending on the hardcoded number of posts to be shown. Also I updated the homebase.sql file to remove my dummy data that doesn't display properly due to incorrect content format, and I added a function that allows one to get the total number of em posts from the database.