Added deletion of a file for file posts.
I added some code to delete a file related to a file upload as a educational material, however this will not quite work until my UploadMaterial branch and this branch are merged into dev. Comment the code to see original functionality, do not forget to use the unlink() php function when deleting files uploaded to the server.